[no ci] fixed some warnings
