+++ external-crates/move/crates/move-package-alt/src/cli/graph.rs
@@ -8,11 +8,9 @@ use crate::{
     errors::PackageResult,
     flavor::Vanilla,
     graph::PackageGraph,
-    package::{EnvironmentName, Package, paths::PackagePath},
+    package::paths::PackagePath,
 };
-use clap::{Command, Parser, Subcommand};
-use petgraph::dot::{Config, Dot};
-use tracing::info;
+use clap::Parser;
 
 /// Build the package
 #[derive(Debug, Clone, Parser)]
