+++ external-crates/move/crates/move-package-alt/src/cli/parse.rs
@@ -5,7 +5,7 @@
 use crate::{
     errors::{FileHandle, PackageResult},
     flavor::Vanilla,
-    package::{lockfile::Lockfiles, manifest::Manifest},
+    package::manifest::Manifest,
     schema::ParsedLockfile,
 };
