+++ external-crates/move/crates/move-package-alt/src/package/manifest.rs
@@ -8,23 +8,16 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use codespan_reporting::{
-    diagnostic::{Diagnostic, Label},
-    term::{
-        self,
-        termcolor::{ColorChoice, StandardStream},
-    },
-};
+use codespan_reporting::diagnostic::{Diagnostic, Label};
 
-use serde::Deserialize;
 use thiserror::Error;
 use tracing::debug;
 
 use crate::{
     dependency::{CombinedDependency, DependencySet},
-    errors::{FileHandle, Files, Located, Location, TheFile},
+    errors::{FileHandle, Location},
     flavor::MoveFlavor,
-    schema::{self, DefaultDependency, ParsedManifest, ReplacementDependency},
+    schema::ParsedManifest,
 };
 
 use super::*;
