+++ external-crates/move/crates/move-package-alt/src/schema/mod.rs
@@ -3,6 +3,7 @@
 mod lockfile;
 mod manifest;
 mod resolver;
+mod sha;
 mod shared;
 
 pub use {lockfile::*, manifest::*, resolver::*, shared::*};
